docs: add schemas for the JSON protocol
This is the first draft of schemas for the JSON protocol KDE Connect
uses to communicate between devices.

These are based on an early draft of JSON Schemas, such that they
should be parsable by any library. In other words, no fancy features
of JSON Schemas are in use here.

The documentation comments assume the text will end up in a MarkDown
document at some point, but again don't contain inline links or
anything other than backticks.

There is a `index.json` included, which groups packets by plugin and
includes a title, description and list of links to implementations of
the plugin. This file has a custom format.